--Menu stuff
git-svn-id: http://svn.jamesgrafton.com/spitfire-maverick/trunk@19 9e128ace-5a45-0410-969d-bf5d6f0db603

+++ spitfire_maverick/spitfire_maverick/header/destructable_object.h
@@ -6,18 +6,18 @@
 class DestructableObject: public GameObject{
 	public:
 		void destructObject();
-		s16 getHealth();
-		void setHealth(s16 health);
+		s32 getHealth();
+		void setHealth(s32 health);
 		u16 getParticleCount();
 		ParticleObject* getParticleSpriteInstance();
 		bool getDestroyed();
 
-		DestructableObject(s32 startx,s32 starty,u16 width,u16 height,s32 heading,s16 vx, s16 vy,SpriteInfo* si,s16 health,s16 particleCount,s16 destroyedGfxRef,ParticleObject* particleSpriteInstance);
+		DestructableObject(s32 startx,s32 starty,u16 width,u16 height,s32 heading,s16 vx, s16 vy,SpriteInfo* si,s32 health,s16 particleCount,s16 destroyedGfxRef,ParticleObject* particleSpriteInstance);
 		DestructableObject(const DestructableObject &object);	//Copy constructor
 		~DestructableObject();
 	private:
 		//Fields
-		s16 health;
+		s32 health;
 		s16 particleCount;
 		s16 destroyedGfxRef;
 		ParticleObject* particleSpriteInstance;
